backend/requirements_enhanced.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python-headless==4.8.1.78
pillow==10.0.1
numpy==1.24.3
requests==2.31.0
google-cloud-vision==3.4.4
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
google-auth==2.23.4
pathlib2==2.3.7 